add functionality

+++ base/views.py
@@ -1,9 +1,61 @@
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
-from .models import Room, Topic
+from django.template import context
+from django.contrib.auth.models import User
+from .models import Message, Room, Topic
 from .forms import RoomForm
 from django.http import JsonResponse
 from django.template.loader import render_to_string
 from django.db.models import Q
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
+
+
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'user does not exist')
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'username or password does not exist')
+
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'error occurred')
+
+    return render(request, 'base/login_register.html', {'form': form})
 
 def home(request):
     is_ajax = request.headers.get('X-Requested-With') == 'XMLHttpRequest'
@@ -22,16 +74,27 @@ def home(request):
         return JsonResponse(dataDict, safe=False)
 
     topics = Topic.objects.all()
+    
     context = {'rooms' : rooms, 'topics' : topics, 'room_count':room_count}
     return render(request, 'base/home.html', context)
 
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room':room}
+    room_messages = room.message_set.all().order_by('-created')
+
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        return redirect('room', pk=room.id)
+
+    context = {'room':room, 'room_messages': room_messages}
     return render(request, 'base/room.html', context)
 
-
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -42,10 +105,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Not your post')
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -54,9 +121,11 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('Not your post')
     context = {'obj': room}
     if request.method == "POST":
         room.delete()
